requirements.txt
fastapi
uvicorn
PyJWT
SQLAlchemy
psycopg2-binary
redis
